+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 ### Added
-- api error exception tag
+- api error exception error
 ### Changed
 - api error exception messages
 ### Fixed